placed login and logout into useAuth
